parsers: reorganize parsers files
